backend/pyproject.toml
[project]
name = "backend"
version = "1.0.0"
description = "Django backend for the full-stack starter kit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "django-extensions>=3.2.3",
    "django-ninja>=1.3.0",
    "django-ninja-extra>=0.22.3",
]
